fix(library): harden experimental pagination (errors, reload memory, escaping)
- Wrap page loads in try/catch so a DB failure logs and retries instead of
  killing the collector that reacts to filter/sort/search changes.
- Rebuild loaded pages in page-sized chunks on refresh so the per-page memory
  ceiling holds while restoring a deep scroll, and track the actually-restored
  page count when data shrank.
- Escape LIKE wildcards in the pushed-down search term (pairs with ESCAPE) so a
  literal term doesn't widen the SQL prefilter.
- Use Long for the load-more sentinel key end-to-end to avoid Int overflow on
  the reset-token multiplier.
- Note the stable-order tiebreaker and mutable-sort-key caveat in the query.

+++ app/src/main/java/eu/kanade/presentation/library/components/LazyLibraryGrid.kt
@@ -36,7 +36,7 @@ internal fun LazyLibraryGrid(
  * even when in-memory filters hid the whole page, so a filtered-out page can't dead-end the
  * category. The key stops changing once the category is exhausted. No-op when [onLoadMore] null.
  */
-internal fun LazyGridScope.loadMoreSentinel(loadKey: Int, onLoadMore: (() -> Unit)?) {
+internal fun LazyGridScope.loadMoreSentinel(loadKey: Long, onLoadMore: (() -> Unit)?) {
     if (onLoadMore == null) return
     item(
         span = { GridItemSpan(maxLineSpan) },
@@ -46,7 +46,7 @@ internal fun LazyGridScope.loadMoreSentinel(loadKey: Int, onLoadMore: (() -> Uni
     }
 }
 
-internal fun LazyListScope.loadMoreSentinel(loadKey: Int, onLoadMore: (() -> Unit)?) {
+internal fun LazyListScope.loadMoreSentinel(loadKey: Long, onLoadMore: (() -> Unit)?) {
     if (onLoadMore == null) return
     item(contentType = "library_load_more") {
         LaunchedEffect(loadKey) { onLoadMore() }
